% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFormula.R
\name{listFormula}
\alias{listFormula}
\title{Create list of Formulas for each voxel}
\usage{
listFormula(x, formula)
}
\arguments{
\item{x}{Index of voxels to be analyzed}

\item{formula}{covariates to be included in the analysis}
}
\description{
This function is internal. 
This function creates list of formulas that will be passed for analysis.
}
\examples{


x <- 1
fm1 <- "~ x1"
formula <- listFormula(x, formula = fm1)
}
\keyword{internal}

